Use Cart boas práticas

+++ src/hooks/useCart.tsx
@@ -34,35 +34,52 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
 
   const addProduct = async (productId: number) => {
     try {
+      const updateCart = [...cart];
+      const productExist = updateCart.find(p => p.id === productId);
+
       const response = await api.get(`/stock/${productId}`);
-      const { amount } = response.data;
+      if(response.status != 200) throw "Erro";
+        const { amount } = response.data;
       if(!amount) throw "Erro";
       const productAmout = cart.find(p => p.id === productId);
       if(productAmout && productAmout.amount + 1 > amount) {
         toast.error('Quantidade solicitada fora de estoque');
         return;
       }
-      if(productAmout) {
-        setCart(cart.map(p => p.id === productId ? { ...p, amount: p.amount + 1 } : p));
+      if(productExist) {
+        productExist.amount += 1;
+        
       }else{
+
         const product =  (await api.get(`/products/${productId}`)).data;
         
-
-        setCart([...cart, { ...product, amount: 1 }]);
+        const newProduct = {
+          ...product,
+          amount : 1
+        }
+        updateCart.push(newProduct);
       }
-      localStorage.setItem('@RocketShoes:cart', JSON.stringify(cart));
+      setCart(updateCart);
+      localStorage.setItem('@RocketShoes:cart', JSON.stringify(updateCart));
 
    } catch {
-      toast.error('Erro ao adicionar o produto no carrinho');
+      toast.error('Erro na adição do produto');
     }
   };
 
   const removeProduct = (productId: number) => {
     try {
-      const product = cart.find(p => p.id === productId);
-      if(!product) throw "Erro";
-      setCart(cart.filter(p => p.id !== productId));
-      localStorage.setItem('@RocketShoes:cart', JSON.stringify(cart));
+      const updateCart = [...cart];
+      const productIndex = updateCart.findIndex(p => p.id === productId);
+
+    
+      if(productIndex >= 0){
+        updateCart.splice(productIndex, 1);
+        setCart(updateCart); 
+        localStorage.setItem('@RocketShoes:cart', JSON.stringify(updateCart));
+      }else{
+        throw Error();
+      }
     } catch {
       toast.error('Erro na remoção do produto');
     }
@@ -82,7 +99,12 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
         toast.error(`Quantidade solicitada fora de estoque`);
         return;
       }
-      setCart(cart.map(p => p.id === productId ? { ...p, amount } : p));
+      const updateCart = [...cart];
+      const productExist = updateCart.find(p => p.id === productId);
+      if(!productExist) return;
+      productExist.amount = amount;
+
+      setCart(updateCart);
       localStorage.setItem('@RocketShoes:cart', JSON.stringify(cart));
 
     } catch {
